add grid component and exchange values
