+++ src/app/app.component.html
@@ -2,47 +2,8 @@
 	<h3>{{ name }}</h3>
 </div>
 
-<div class="row d-flex justify-content-center">
-	<div class="col-md-3">
-		<div *ngFor='let row of data; let i=index' class="d-flex justify-content-center">
-			<div *ngFor='let cell of row; let j=index' class="d-flex justify-content-center">
-				<div [ngClass]="(i>2&&i<6 || j>2&&j<6)?'ls-cover':'mr-cover'">
-					<div [ngClass]="(i>2&&i<6 && j>2&&j<6)?'mr-cover':''">
-						<button
-              [disabled]="!cell.s"
-              id="{{i+1}}-{{j+1}}"
-              class="square"
-              [ngClass]="cell.s ? 'enable':'disable' "
-              (click)="open(i,j,content)"
-            >
-             <span> {{cell.v}} </span>
-            </button>
-					</div>
-				</div>
-			</div>
-		</div>
-		<div class="difficulty d-flex justify-content-center">
-			<button class="levels"
-            (click)="setDifficulty('easy')">Easy</button>
-			<button class="levels"
-            (click)="setDifficulty('medium')">Medium</button>
-			<button class="levels"
-            (click)="setDifficulty('advanced')">Advanced</button>
-		</div>
-	</div>
-</div>
-
-
-<ng-template #content let-modal>
-	<!-- <div class="modal-header">
-    <button type="button" class="close" aria-label="Close" (click)="modal.dismiss('Cross click')">
-      <span aria-hidden="true">&times;</span>
-    </button>
-  </div> -->
-	<!-- <div class="modal-body"> -->
-
-	<app-num-picker (click)="modal.dismiss('Cross click')" (messageEvent)="recieveValue($event)">
-	</app-num-picker>
-
-	<!-- </div> -->
-</ng-template>
+<app-grid 
+    [init]="difficultyCells"
+    (difficulty)="setDifficulty($event)"
+    >
+</app-grid>
